starting over. Adding everything into Game.jsx
